requirements.txt
numpy
Flask
sklearn
pandas
PyYAML
evidently
dill
matplotlib
seaborn
plotly
chart-studio
pymongo[srv]
boto3
pytest
python-dotenv
